third_party/WebKit/LayoutTests/platform/linux/http/tests/devtools/sources/debugger-ui/source-frame-count-expected.txt

Tests that scripts panel does not create too many source frames.


Running: testSourceFramesCount
Reloading page...
Page reloaded.
Less than 3 frames opened
Visible view: script5.js
